asteres/requirements.txt
Flask==1.0.2
Flask-Cors==2.1.2
Flask-WTF==0.14.2
futures==3.0.5
html5lib==0.999999999
httplib2==0.10.3
Jinja2>=2.10.1
MarkupSafe==0.23
matplotlib==1.5.3
oauth2client==4.0.0
oauthlib==2.0.1
requests>=2.20.0
requests-oauthlib==0.8.0
werkzeug>=0.15.3
wsgi-request-logger==0.4.6
WTForms==2.1
pandas==0.18.1
unidecode==0.4.20
PyGithub==1.47
